% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rechordr.R
\name{rechordr}
\alias{rechordr}
\title{Play or save musical chords}
\usage{
rechordr(pitch, duration, tempo = 120, sample_rate = 44100, fade_duration = 0.02, output_path = NULL)
}
\arguments{
\item{pitch}{a character vector or list of notes using the Latin alphabet. Use a list to play multiple notes simultaneously, placing all notes to be played together in a character vector. An element with no characters results in silence. Append:
\itemize{
\item \verb{#} for sharp semitones
\item \code{b} for flat semitones
\item octave number, by default octave 4
}}

\item{duration}{a numeric vector of each note's duration, in beats. Must be same length as pitch.}

\item{tempo}{integer, tempo to be played in beats per minute.}

\item{sample_rate}{integer, number of samples per second (Hz).}

\item{fade_duration}{number of seconds to fade each note in and out. Can be used to avoid audible pops at the beginning and end of each note. Cannot exceed half of the shortest note's duration.}

\item{output_path}{path to location where wave file will be written. If \code{NULL}, music is played in the console.}
}
\value{
Music played in R console, or saved wave file.
}
\description{
Play or save musical chords
}
\examples{
rechordr(pitch = list(c("A3", "C4", "F4"), c("G3", "C4", "E4"), c("G3", "C4", "E4"), c("F3", "Bb3", "D4"), "", c("A3", "C4", "F4"), c("G3", "C4", "E4"), c("G3", "C4", "E4"), c("A3", "C4", "F4")),
         duration = c(1, .75, .75, .5, 1, 1, .75, .75, .5))

rechordr(pitch = list(c("A3", "C4", "F4"), c("G3", "C4", "E4"), c("G3", "C4", "E4"), c("F3", "Bb3", "D4"), "", c("A3", "C4", "F4"), c("G3", "C4", "E4"), c("G3", "C4", "E4"), c("A3", "C4", "F4")),
         duration = c(1, .75, .75, .5, 1, 1, .75, .75, .5),
         output_path = "You Can Call Me Al.wav")

}
\author{
Ariel Fridman
}
